+++ js/utils.js
@@ -1,18 +1,12 @@
 // ************ Big Feature related ************
 
 function respecBuyables(layer) {
-	if (!layers[layer].buyables) {
-		return;
-	}
-	if (!layers[layer].buyables.respec) {
-		return;
-	}
-	if (!confirm("Are you sure you want to respec? This will force you to do a \"" + (tmp[layer].name ? tmp[layer].name : layer) + "\" reset as well!")) {
-		return;
-	}
-	run(layers[layer].buyables.respec, layers[layer].buyables);
-	updateBuyableTemp(layer);
-	document.activeElement.blur();
+	if (!layers[layer].buyables) return
+	if (!layers[layer].buyables.respec) return
+	if (!player[layer].noRespecConfirm && !confirm(tmp[layer].buyables.respecMessage || "Are you sure you want to respec? This will force you to do a \"" + (tmp[layer].name ? tmp[layer].name : layer) + "\" reset as well!")) return
+	run(layers[layer].buyables.respec, layers[layer].buyables)
+	updateBuyableTemp(layer)
+	document.activeElement.blur()
 }
 
 function canAffordUpgrade(layer, id) {
@@ -278,17 +272,24 @@ function layOver(obj1, obj2) {
 }
 
 function prestigeNotify(layer) {
-	if (layers[layer].prestigeNotify) {
-		return layers[layer].prestigeNotify();
-	} else if (tmp[layer].autoPrestige || tmp[layer].passiveGeneration) {
-		return false;
-	} else if (tmp[layer].type === "static") {
-		return tmp[layer].canReset;
-	} else if (tmp[layer].type === "normal") {
-		return (tmp[layer].canReset && (tmp[layer].resetGain.gte(player[layer].points.div(10))));
-	} else {
-		return false;
+	if (layers[layer].prestigeNotify) return layers[layer].prestigeNotify()
+	
+	if (isPlainObject(tmp[layer].tabFormat)) {
+		for (subtab in tmp[layer].tabFormat){
+			if (subtabResetNotify(layer, 'mainTabs', subtab))
+				return true
+		}
 	}
+	for (family in tmp[layer].microtabs) {
+		for (subtab in tmp[layer].microtabs[family]){
+			if (subtabResetNotify(layer, family, subtab))
+				return true
+		}
+	}
+	if (tmp[layer].autoPrestige || tmp[layer].passiveGeneration) return false
+	else if (tmp[layer].type == "static") return tmp[layer].canReset
+	else if (tmp[layer].type == "normal") return (tmp[layer].canReset && (tmp[layer].resetGain.gte(player[layer].points.div(10))))
+	else return false
 }
 
 function notifyLayer(name) {
@@ -314,17 +315,11 @@ function subtabShouldNotify(layer, family, id) {
 }
 
 function subtabResetNotify(layer, family, id) {
-	let subtab;
-	if (family === "mainTabs") {
-		subtab = tmp[layer].tabFormat[id];
-	} else {
-		subtab = tmp[layer].microtabs[family][id];
-	}
-	if (subtab.embedLayer) {
-		return tmp[subtab.embedLayer].prestigeNotify;
-	} else {
-		return false;
-	}
+	let subtab = {}
+	if (family == "mainTabs") subtab = tmp[layer].tabFormat[id]
+	else subtab = tmp[layer].microtabs[family][id]
+	if (subtab.embedLayer) return tmp[subtab.embedLayer].prestigeNotify
+	else return subtab.prestigeNotify
 }
 
 function nodeShown(layer) {
@@ -414,35 +409,34 @@ function addTime(diff, layer) {
 	}
 }
 
-let onFocused;
-document.onkeydown = function (e) {
-	if (player === undefined) {
-		return;
-	}
-	if (gameEnded && !player.keepGoing) {
-		return;
-	}
+shiftDown = false
+ctrlDown = false
 
-	let ctrlDown = e.ctrlKey;
-	let key = e.key;
-	if (ctrlDown) {
-		key = "ctrl+" + key;
-	}
-	if (onFocused) {
-		return;
-	}
-	if (ctrlDown && hotkeys[key]) {
-		e.preventDefault();
-	}
+document.onkeydown = function (e) {
+	if (player === undefined) return;
+	if (gameEnded && !player.keepGoing) return;
+	shiftDown = e.shiftKey
+	ctrlDown = e.ctrlKey
+	let key = e.key
+	if (ctrlDown) key = "ctrl+" + key
+	if (onFocused) return
+	if (ctrlDown && hotkeys[key]) e.preventDefault()
 	if (hotkeys[key]) {
-		if (player[hotkeys[key].layer].unlocked) {
-			hotkeys[key].onPress();
-		}
+		let k = hotkeys[key]
+		console.log(tmp[k.layer].hotkeys)
+		if (player[k.layer].unlocked && tmp[k.layer].hotkeys[k.id].unlocked)
+			k.onPress()
 	}
 };
 
 onFocused = false;
 
+document.onkeyup = function (e) {
+	shiftDown = e.shiftKey
+	ctrlDown = e.ctrlKey
+}
+
+var onFocused = false
 function focused(x) {
 	onFocused = x;
 }
@@ -472,6 +466,14 @@ function isPlainObject(obj) {
 
 document.title = modInfo.name;
 
+// Converts a string value to whatever it's supposed to be
+function toValue(value, oldValue) {
+	if (oldValue instanceof Decimal)
+		return new Decimal (value)
+	else if (!isNaN(oldValue))
+		return value.toNumber()
+	else return value
+}
 
 // Variables that must be defined to display popups
 const activePopups = [];