+++ js/utils/NumberFormatting.js
@@ -6,8 +6,8 @@ function exponentialFormat(num, precision, mantissa = true) {
 		m = new Decimal(1);
 		e = e.add(1);
 	}
-	e = commaFormat(e);
-	if (mantissa) {
+	e = (e.gte(1e9) ? format(e, 1) : (e.gte(10000) ? commaFormat(e, 0) : e.toStringWithDecimalPlaces(0)))
+    if (mantissa) {
 		return m.toStringWithDecimalPlaces(precision)+"e"+e;
 	} else {
 		return "e"+e;