+++ js/utils/save.js
@@ -80,8 +80,14 @@ function getStartLayerData(layer) {
 	if (layerdata.resetTime === undefined) {
 		layerdata.resetTime = 0;
 	}
+	if (layerdata.forceTooltip === undefined) {
+		layerdata.forceTooltip = false;
+	}
 
 	layerdata.buyables = getStartBuyables(layer);
+	if (layerdata.noRespecConfirm === undefined) {
+		layerdata.noRespecConfirm = false;
+	}
 	if (layerdata.clickables === undefined) {
 		layerdata.clickables = getStartClickables(layer);
 	}
@@ -190,7 +196,7 @@ function load() {
 	if (get === null || get === undefined) {
 		player = getStartPlayer();
 	} else {
-		player = Object.assign(getStartPlayer(), JSON.parse(atob(get)));
+		player = Object.assign(getStartPlayer(), JSON.parse(decodeURIComponent(escape(atob(get)))));
 	}
 	fixSave();